object_localization/requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==20.3.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
Bottleneck==1.2.1
cephfs==2.0.0
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
click==8.1.8
ClusterShell==1.8.3
cryptography==3.3.2
csbdeep==0.8.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docker-pycreds==0.4.0
einops==0.8.1
et-xmlfile==1.0.1
filelock==3.0.12
GDAL==3.2.2
git-cola==3.9
gitdb==4.0.12
GitPython==3.1.44
html5lib==1.1
httplib2==0.18.1
hydra-core==1.3.2
idna==2.10
imageio==2.37.0
importlib-metadata==1.6.0
iniconfig==0.0.0
invoke==1.4.1
ipython==7.20.0
ipython-genutils==0.2.0
jdcal==1.0
jedi==0.18.0
Jinja2==2.11.3
kiwisolver==1.3.1
lazy-loader==0.4
libvirt-python==7.0.0
llvmlite==0.35.0
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mock==4.0.3
more-itertools==4.2.0
mysqlclient==1.4.4
networkx==3.2.1
numba==0.52.0
numexpr==2.7.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
odfpy==1.4.1
olefile==0.46
omegaconf==2.3.0
openpyxl==3.0.3
packaging==24.2
pandas==1.1.5
paramiko==2.7.2
parso==0.8.1
pathtools==0.1.2
pbr==5.5.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
pluggy==0.13.0
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.14
protobuf==4.25.6
psutil==7.0.0
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
pydot==1.4.2
Pygments==2.7.1
PyGObject==3.38.0
pygraphviz==1.7
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pytest==6.0.2
python-dateutil==2.8.1
pytz==2021.1
pywavelets==1.6.0
PyYAML==5.3.1
QtPy==1.9.0
rados==2.0.0
rbd==2.0.0
requests==2.25.1
scikit-image==0.21.0
scipy==1.11.3
sentry-sdk==2.25.0
setproctitle==1.3.5
six==1.16.0
smmap==5.0.2
soupsieve==2.2.1
tables==3.6.1
tifffile==2024.8.30
toml==0.10.1
torch==1.13.0
tqdm==4.67.1
traitlets==5.0.5
typing-extensions==4.13.0
urllib3==2.3.0
virtualenv==20.4.0+ds
wandb==0.15.12
wcwidth==0.1.9
webencodings==0.5.1
xlwt==1.3.0
zipp==1.0.0
